Remove upcast-selection-converters.js.
